refactor: :recycle: Extract `-e` into function
